Change member priority from int to float64 for decimal priorities
MongoDB accepts decimal priority values (e.g., 0.1 for analytical nodes).
Update ConfigMember.Priority and MemberOverride.Priority to float64,
change schema from TypeInt to TypeFloat, add toFloat64 helper for
Terraform state coercion. Update all tests for float64 comparisons.
Refactor shard_config examples to use locals-based merge pattern with
analytical/online node type defaults.

+++ mongodb/replica_set_types.go
@@ -49,7 +49,7 @@ type ConfigMember struct {
 	ArbiterOnly  *bool       `bson:"arbiterOnly,omitempty" json:"arbiterOnly,omitempty"`
 	BuildIndexes *bool       `bson:"buildIndexes,omitempty" json:"buildIndexes,omitempty"`
 	Hidden       *bool       `bson:"hidden,omitempty" json:"hidden,omitempty"`
-	Priority     int         `bson:"priority,omitempty" json:"priority,omitempty"`
+	Priority     float64     `bson:"priority,omitempty" json:"priority,omitempty"`
 	Tags         ReplsetTags `bson:"tags,omitempty" json:"tags,omitempty"`
 	SlaveDelay   *int64      `bson:"slaveDelay,omitempty" json:"slaveDelay,omitempty"`
 	Votes        *int        `bson:"votes,omitempty" json:"votes,omitempty"`
